Fixed all override errors in visualization

+++ moveit_ros/visualization/motion_planning_rviz_plugin/include/moveit/motion_planning_rviz_plugin/interactive_marker_display.h
@@ -82,7 +82,7 @@ public:
   }
 
   // Overrides from Display
-  void update(double wall_dt, double ros_dt) override;
+  void update(float wall_dt, float ros_dt) override;
 
   void reset() override;
